backend/requirements.txt

fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
jinja2

# pip install -r backend/requirements.txt